Added .bsdesign file, added option to download resume
Added the .bsdesign file containing the new design. It has been renamed to "Shakeel Khan", which will appear as the name of the site on the browser tab (there is probably a more formal way to go about this, but this is will do for now). Also added buttons to download a copy of the resume to the footer of each page, and updated the link to the GitHub repo in the footer on each page. Updated resume as well.